agents/pyproject.toml
# ============================================
# Galderma TrackWise AI Autopilot Demo
# Agents - 9 Strands Agents on AgentCore
# ============================================
#
# All agents use AWS Strands Agents SDK and deploy to
# AWS Bedrock AgentCore Runtime.
#
# Agent List:
# 1. Observer (Haiku) - Orchestrator
# 2. Case Understanding (Haiku)
# 3. Recurring Detector (Haiku)
# 4. Compliance Guardian (OPUS) - Critical
# 5. Resolution Composer (OPUS) - Critical
# 6. Inquiry Bridge (Haiku)
# 7. Writeback (Haiku)
# 8. Memory Curator (Haiku)
# 9. CSV Pack (Haiku)
# ============================================

[project]
name = "galderma-trackwise-agents"
version = "0.1.0"
description = "9 Strands Agents for Galderma TrackWise AI Autopilot"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "Galderma Demo Team", email = "demo@galderma.com" }
]

dependencies = [
    # AWS Strands Agents (MANDATORY - Core SDK)
    "strands-agents>=0.1.0",
    # Strands A2A Protocol (MANDATORY for inter-agent communication)
    "strands-agents[a2a]>=0.1.0",
    # Strands Built-in Tools
    "strands-agents-tools>=0.1.0",
    # A2A Client for agent discovery
    "strands-agents-tools[a2a_client]>=0.1.0",
    # Bedrock AgentCore (MANDATORY for deployment)
    "bedrock-agentcore>=0.1.0",
    # Bedrock AgentCore Starter Toolkit (for CLI: agentcore launch)
    "bedrock-agentcore-starter-toolkit>=0.1.0",
    # Pydantic v2 (structured outputs)
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    # AWS SDK (for AgentCore Memory, DynamoDB, S3)
    "boto3>=1.34.0",
    # HTTP client (for A2A communication fallback)
    "httpx>=0.27.0",
    # Async support
    "anyio>=4.3.0",
    # Date/time handling
    "python-dateutil>=2.9.0",
    # UUID generation
    "ulid-py>=1.1.0",
    # JSON processing
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    # Linting
    "ruff>=0.3.0",
    # Type checking
    "mypy>=1.8.0",
    "boto3-stubs[bedrock,bedrock-agentcore,dynamodb,s3,logs]>=1.34.0",
    # Mocking
    "moto[dynamodb,s3]>=5.0.0",
    "respx>=0.20.0",
    # Local agent testing
    "pytest-httpx>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "observer", "case_understanding", "recurring_detector",
            "compliance_guardian", "resolution_composer", "inquiry_bridge",
            "writeback", "memory_curator", "csv_pack"]

# ============================================
# Ruff Configuration (Linting & Formatting)
# ============================================
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["shared", "observer", "case_understanding", "recurring_detector",
       "compliance_guardian", "resolution_composer", "inquiry_bridge",
       "writeback", "memory_curator", "csv_pack", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in argument defaults
    "B904",   # raise without from in except
]

[tool.ruff.lint.isort]
known-first-party = ["shared"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# ============================================
# Pytest Configuration
# ============================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests with AgentCore",
    "slow: Slow tests",
    "a2a: Agent-to-Agent protocol tests",
]

# ============================================
# MyPy Configuration
# ============================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "strands.*",
    "strands_agents.*",
    "strands_tools.*",
    "bedrock_agentcore.*",
    "moto.*",
]
ignore_missing_imports = true

# ============================================
# Coverage Configuration
# ============================================
[tool.coverage.run]
source = ["shared", "observer", "case_understanding", "recurring_detector",
          "compliance_guardian", "resolution_composer", "inquiry_bridge",
          "writeback", "memory_curator", "csv_pack"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
